fix: increase mask size to 64 bits

+++ crates/interpreter/src/bit_field.rs
@@ -14,19 +14,19 @@
 
 use crate::error::*;
 
-pub fn into_signed_field(mask: u32, value: i32) -> Result<u32, Error> {
-    into_field(mask, value.wrapping_add(offset(mask)) as u32)
+pub fn into_signed_field(mask: u64, value: i32) -> Result<u64, Error> {
+    into_field(mask, value.wrapping_add(offset(mask)) as u64)
 }
 
-pub fn from_signed_field(mask: u32, field: u32) -> i32 {
+pub fn from_signed_field(mask: u64, field: u64) -> i32 {
     from_field(mask, field) as i32 - offset(mask)
 }
 
-fn offset(mask: u32) -> i32 {
+fn offset(mask: u64) -> i32 {
     1 << (mask.count_ones() - 1)
 }
 
-pub fn into_field(mask: u32, value: u32) -> Result<u32, Error> {
+pub fn into_field(mask: u64, value: u64) -> Result<u64, Error> {
     let field = (value << mask.trailing_zeros()) & mask;
     if from_field(mask, field) != value {
         return Err(unsupported(if_debug!(Unsupported::SideTable)));
@@ -34,7 +34,7 @@ pub fn into_field(mask: u32, value: u32) -> Result<u32, Error> {
     Ok(field)
 }
 
-pub fn from_field(mask: u32, field: u32) -> u32 {
+pub fn from_field(mask: u64, field: u64) -> u64 {
     (field & mask) >> mask.trailing_zeros()
 }
 
@@ -80,7 +80,7 @@ mod tests {
     #[test]
     fn unsigned_field_error() {
         let mask = 0b111000;
-        for value in [8, u32::MAX] {
+        for value in [8, u64::MAX] {
             assert!(matches!(into_field(mask, value), Err(Error::Unsupported(_))));
         }
     }
